# src/openapi.yaml
openapi: 3.1.0
info:
  title: MCP REST Bridge
  version: 1.0.0
  description: >
    A self-hosted MCP gateway that bridges MCP clients to backend REST APIs and
    other MCP servers, bidirectionally. REST-capable services register
    themselves (via OpenAPI/Swagger auto-discovery, a cURL/Postman import, or a
    manual tool list) and the bridge exposes their tools to any MCP client
    (e.g. Claude Desktop, Cursor, or a custom agent) via the Streamable HTTP
    MCP transport; existing MCP servers can likewise be registered as
    upstreams and re-exposed. Every call is proxied through a uniform guard
    pipeline (SSRF check, guardrails, per-tool policy, circuit breaker,
    response sanitizing, audit) documented in this admin API — this
    specification covers that management surface, not the dynamic per-client
    tool endpoints it provisions.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /admin-api/consumers:
    get:
      operationId: listConsumers
      summary: List consumers, with this-month usage
      tags: [Consumers]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      responses:
        "200":
          description: All consumers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/Consumer"
                        - type: object
                          properties:
                            usedThisMonth: { type: integer }
    post:
      operationId: createConsumer
      summary: Create a new consumer
      description: A consumer groups one or more MCP keys under a shared monthly call quota and/or per-end-user rate limit. Requires the admin role for session callers.
      tags: [Consumers]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 128 }
                monthlyQuota: { type: [integer, "null"], description: "Positive integer, or null for unlimited." }
                endUserRateLimitPerMin:
                  { type: [integer, "null"], description: "Positive integer, or null to disable." }
      responses:
        "201":
          description: Created.
          content: { application/json: { schema: { $ref: "#/components/schemas/Consumer" } } }
        "400":
          description: name missing/too long, or an invalid quota/rate-limit value.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: A consumer with that name already exists.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/consumers/{id}:
    patch:
      operationId: updateConsumer
      summary: Update a consumer's name, quota, and/or rate limit
      description: Requires the admin role for session callers.
      tags: [Consumers]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                monthlyQuota: { type: [integer, "null"] }
                endUserRateLimitPerMin: { type: [integer, "null"] }
      responses:
        "200":
          description: Updated.
          content: { application/json: { schema: { $ref: "#/components/schemas/Consumer" } } }
        "400":
          description: Invalid field value.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such consumer.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: Another consumer already has that name.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    delete:
      operationId: deleteConsumer
      summary: Delete a consumer
      description: Requires the admin role for session callers.
      tags: [Consumers]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted.
          content:
            {
              application/json:
                {
                  schema:
                    { type: object, properties: { status: { type: string, example: deleted }, id: { type: integer } } },
                },
            }
        "404":
          description: No such consumer.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
  /admin-api/teams:
    get:
      operationId: listTeams
      summary: List teams
      description: Any authenticated admin caller may read the team list (needed for assignment dropdowns), regardless of role or tenancy scope.
      tags: [Teams]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      responses:
        "200":
          description: All teams.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties: { items: { type: array, items: { $ref: "#/components/schemas/Team" } } },
                    },
                },
            }
    post:
      operationId: createTeam
      summary: Create a new team
      description: Requires a super-admin (an admin-role session with no team assignment). Bearer callers always pass.
      tags: [Teams]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, pattern: "^[a-zA-Z0-9][a-zA-Z0-9 _-]{0,62}$" }
      responses:
        "201":
          description: Created.
          content: { application/json: { schema: { $ref: "#/components/schemas/Team" } } }
        "400":
          description: Invalid name.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "403":
          description: Caller is not a super-admin.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: A team with that name already exists.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/teams/{id}:
    get:
      operationId: getTeam
      summary: Get a single team
      description: Any authenticated admin caller may read a team's detail, same tier as the list endpoint.
      tags: [Teams]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: The team.
          content: { application/json: { schema: { $ref: "#/components/schemas/Team" } } }
        "400":
          description: id is not an integer.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such team.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    delete:
      operationId: deleteTeam
      summary: Delete a team
      description: Requires a super-admin. Clients/users owned by the team are not deleted — check how deleteTeam() reassigns them before relying on this in a scripted flow.
      tags: [Teams]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted.
          content:
            {
              application/json:
                {
                  schema:
                    { type: object, properties: { status: { type: string, example: deleted }, id: { type: integer } } },
                },
            }
        "403":
          description: Caller is not a super-admin.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such team.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/clients/{name}/team:
    put:
      operationId: setClientTeam
      summary: Assign (or clear) a client's owning team
      description: Requires a super-admin.
      tags: [Teams]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId]
              properties:
                teamId:
                  {
                    type: [integer, "null"],
                    description: null clears ownership (client becomes visible only to super-admins).,
                  }
      responses:
        "200":
          description: Updated.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties:
                        { status: { type: string }, name: { type: string }, teamId: { type: [integer, "null"] } },
                    },
                },
            }
        "400":
          description: teamId missing or not a number/null.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "403":
          description: Caller is not a super-admin.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such client or team.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/users/{username}/team:
    put:
      operationId: setUserTeam
      summary: Assign (or clear) an admin user's team membership
      description: A user with teamId null is a super-admin over tenancy — sees and mutates every team's clients, and can reassign ownership. Requires a super-admin.
      tags: [Teams]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: username, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId]
              properties:
                teamId: { type: [integer, "null"] }
      responses:
        "200":
          description: Updated.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties:
                        { status: { type: string }, username: { type: string }, teamId: { type: [integer, "null"] } },
                    },
                },
            }
        "400":
          description: teamId missing or not a number/null.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "403":
          description: Caller is not a super-admin.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such user or team.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: LAST_SUPERADMIN_PROTECTED — cannot assign a team to the last active super-admin (admin role, no team).
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }

  # ── Traces ────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      description: >
        API key for administrative endpoints (registration, introspection).
        Set via ADMIN_API_KEYS env var. Disabled when AUTH_DISABLED=true.
    McpAuth:
      type: http
      scheme: bearer
      description: >
        API key for the backend-tool MCP transport endpoints
        (/mcp/{clientName}, /mcp-custom/{bundleName}). Set via MCP_API_KEYS
        env var, or a managed key (see MCP Keys). Disabled when
        AUTH_DISABLED=true or when no keys are configured at all.
    RootMcpAuth:
      type: http
      scheme: bearer
      description: >
        Credential for the /mcp system control-plane. Either the env admin
        Bearer key (ADMIN_API_KEYS) or a managed MCP key with `adminRole`
        set (see MCP Keys). Unlike McpAuth, there is no "no auth material
        configured => allow all" fallback — a caller must resolve to a real
        system role or the request is rejected outright, even on a fresh
        install with zero keys minted. Still bypassed entirely when
        AUTH_DISABLED=true (the same global dev/test escape hatch every
        other auth path in this API respects).
    SessionAuth:
      type: apiKey
      in: cookie
      name: __Host-mcp_admin_session
      description: >
        Admin-UI session cookie, issued by POST /admin-api/auth/login.
        Accepted anywhere AdminAuth is accepted. On mutating requests
        (anything but GET/HEAD/OPTIONS) an `X-CSRF-Token` header matching the
        `__Host-mcp_admin_csrf` cookie value is also required — OpenAPI 3.1
        has no first-class way to express that pairing, so it's documented
        here in prose, mirroring how RegistrationPayload's tools/openapi_url
        mutual-exclusivity is documented rather than schema-encoded.
        The `__Host-` prefix (and therefore this exact name) only applies
        when SESSION_COOKIE_SECURE=true (the default) — it mandates the
...
  schemas:
...
    Consumer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        monthlyQuota: { type: [integer, "null"], description: "Positive integer, or null for unlimited." }
        endUserRateLimitPerMin: { type: [integer, "null"], description: "Positive integer, or null to disable." }
        createdAt: { type: integer, description: Unix ms epoch }
        updatedAt: { type: integer, description: Unix ms epoch }
        createdBy: { type: [string, "null"] }

    # -------------------------------------------------------------------------
    # Policies
    # -------------------------------------------------------------------------
...
    Team:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        createdAt: { type: integer, description: Unix ms epoch }
        createdBy: { type: [string, "null"] }

    # -------------------------------------------------------------------------
    # Traces
    # -------------------------------------------------------------------------
...
    ApiError:
      type: object
      description: >
        Standard error envelope returned by the bridge's REST endpoints (not
        MCP transport endpoints, which use JSON-RPC 2.0 error objects instead).
      required:
        - error
      properties:
        error:
          type: object
          description: Container for the structured error detail.
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Machine-readable error code in SCREAMING_SNAKE_CASE.
                Known values: VALIDATION_ERROR, CLIENT_NOT_FOUND, DISCOVERY_ERROR,
                UNAUTHORIZED, FORBIDDEN, RATE_LIMITED.
              example: VALIDATION_ERROR
            message:
              type: string
              description: >
                Human-readable description of what went wrong, suitable for
                displaying in a developer console or log.
              example: "Missing required fields: name, health_url"
            retry_after:
              type: integer
              description: >
                Seconds the caller should wait before retrying. Present only on
                a 429 RATE_LIMITED response (also echoed in the Retry-After
                header); omitted on every other error.
              example: 45
            request_id:
              type: [string, "null"]
              description: >
                Correlation id for this request (from res.locals.requestId),
                echoed back so a caller can quote it when reporting an issue.
                Present on almost every error response in the admin API;
                absent (or omitted) on the handful of handlers that don't
                thread it through (notably GET /clients and GET
                /clients/{name}/tools's error bodies, and /metrics's disabled
                response).
              example: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"
      example:
        error:
          code: VALIDATION_ERROR
          message: "Missing required fields: name, health_url"
          request_id: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"

    # -------------------------------------------------------------------------
    # JsonRpcMessage  (internal — MCP transport)
    # -------------------------------------------------------------------------
...