# src/routes/admin/teams.ts
import { Router } from "express";
import type { Request, Response } from "express";
import { requireSuperAdmin } from "../../middleware/authz.js";
import { recordAudit, actorFromRequest } from "../../admin/audit/audit.js";
import { listTeams, getTeam, createTeam, deleteTeam, setClientTeam, setUserTeam } from "../../admin/entities/teams.js";
import { findUserByUsername, countActiveTeamlessAdmins } from "../../security/user-store.js";
import { sendError, validationError, notFound, bodyOf } from "../http-errors.js";

export const teamRoutes = Router();

// Any admin may read the team list (needed for assignment dropdowns).
teamRoutes.get("/teams", (_req: Request, res: Response) => {
  res.status(200).json({ items: listTeams() });
});

// Any admin may read a single team's detail (same auth tier as the list above).
teamRoutes.get("/teams/:id", (req: Request<{ id: string }>, res: Response) => {
  const id = Number(req.params.id);
  if (!Number.isInteger(id)) {
    validationError(res, "id must be an integer");
    return;
  }
  const team = getTeam(id);
  if (!team) {
    notFound(res, "TEAM_NOT_FOUND", "Team not found");
    return;
  }
  res.status(200).json(team);
});

teamRoutes.post("/teams", requireSuperAdmin, (req: Request, res: Response) => {
  const body = bodyOf(req);
  const name = typeof body.name === "string" ? body.name.trim() : "";
  if (!name) {
    validationError(res, "name is required");
    return;
  }
  const result = createTeam(name, actorFromRequest(req));
  if (result === "INVALID_NAME") {
    sendError(res, 400, "INVALID_NAME", "Team name must be 1-63 chars: letters, digits, space, - or _");
    return;
  }
  if (result === "ALREADY_EXISTS") {
    sendError(res, 409, "ALREADY_EXISTS", "A team with that name already exists");
    return;
  }
  recordAudit(actorFromRequest(req), "team.create", `team:${result.id}`, { name: result.name });
  res.status(201).json(result);
});

teamRoutes.delete("/teams/:id", requireSuperAdmin, (req: Request<{ id: string }>, res: Response) => {
  const ok = deleteTeam(Number(req.params.id));
  if (!ok) {
    notFound(res, "TEAM_NOT_FOUND", "Team not found");
    return;
  }
  recordAudit(actorFromRequest(req), "team.delete", `team:${req.params.id}`);
  res.status(200).json({ status: "deleted", id: Number(req.params.id) });
});

// Assign (or clear) a client's owning team.
teamRoutes.put("/clients/:name/team", requireSuperAdmin, (req: Request<{ name: string }>, res: Response) => {
  const body = bodyOf(req);
  const teamId = body.teamId === null ? null : typeof body.teamId === "number" ? body.teamId : undefined;
  if (teamId === undefined) {
    validationError(res, "teamId must be a number or null");
    return;
  }
  const ok = setClientTeam(req.params.name, teamId);
  if (!ok) {
    notFound(res, "NOT_FOUND", "Client or team not found");
    return;
  }
  recordAudit(actorFromRequest(req), "client.team.set", req.params.name, { teamId });
  res.status(200).json({ status: "updated", name: req.params.name, teamId });
});

// Assign (or clear) a user's team membership.
teamRoutes.put("/users/:username/team", requireSuperAdmin, (req: Request<{ username: string }>, res: Response) => {
  const body = bodyOf(req);
  const teamId = body.teamId === null ? null : typeof body.teamId === "number" ? body.teamId : undefined;
  if (teamId === undefined) {
    validationError(res, "teamId must be a number or null");
    return;
  }
  // Guard against assigning away the last active teamless (super-admin)
  // user: every super-admin-only action (team CRUD, ownership assignment,
  // SSO config...) requires role==='admin' && team_id===null, and there is
  // no session-based path back once every admin has a team — same hazard
  // class routes/admin/users.ts already guards for role/active-state via
  // countActiveAdmins().
  if (teamId !== null) {
    const target = findUserByUsername(req.params.username);
    if (
      target &&
      target.role === "admin" &&
      target.isActive &&
      target.teamId === null &&
      countActiveTeamlessAdmins() <= 1
    ) {
      sendError(
        res,
        409,
        "LAST_SUPERADMIN_PROTECTED",
        "Cannot assign a team to the last active super-admin (admin role, no team)",
      );
      return;
    }
  }
  const ok = setUserTeam(req.params.username, teamId);
  if (!ok) {
    notFound(res, "NOT_FOUND", "User or team not found");
    return;
  }
  recordAudit(actorFromRequest(req), "user.team.set", req.params.username, { teamId });
  res.status(200).json({ status: "updated", username: req.params.username, teamId });
});